make the code a little more readable
