+++ pages/doc-components/index.js
@@ -4,5 +4,5 @@ export {default as SideNav} from './SideNav'
 export {default as IndexHero} from './IndexHero'
 export {default as BoxShadow} from './BoxShadow'
 export {default as PropsList} from './PropsList'
-export {default as LiveCode} from './LiveCode'
+export {default as CodeContainer} from './CodeContainer'
 export * from './_utils'
